+++ src/org/andengine/examples/TexturePackerExample.java
@@ -75,8 +75,8 @@ public class TexturePackerExample extends BaseExample {
 
 		TextureRegion faceTextureRegion = this.mSpritesheetTexturePackTextureRegionLibrary.get(TexturePackerExampleSpritesheet.FACE_BOX_ID);
 		/* Calculate the coordinates for the face, so its centered on the camera. */
-		final int centerX = (CAMERA_WIDTH - faceTextureRegion.getWidth()) / 2;
-		final int centerY = (CAMERA_HEIGHT - faceTextureRegion.getHeight()) / 2;
+		final float centerX = (CAMERA_WIDTH - faceTextureRegion.getWidth()) / 2;
+		final float centerY = (CAMERA_HEIGHT - faceTextureRegion.getHeight()) / 2;
 
 		/* Create the face and add it to the scene. */
 		Sprite entity = new Sprite(centerX, centerY, faceTextureRegion, this.getVertexBufferObjectManager());
